generator/requirements.txt
kafka-python
avro-python3==1.8.2; python_version == '3.7'
confluent-kafka==1.1.0; python_version == '3.7'
confluent-kafka[avro]==1.1.0; python_version < '3.7'
